+++ src/store/red-packet.ts
@@ -1,3 +1,7 @@
 export const noWinRef = {
   value: 0,
 };
+
+export const realRisk = {
+  value: false,
+};
